Moved component files into new components directory
